Add support for Enclustra Mars ZX3 board as a BMC
This board contains a Xilinx Zynq 7020 which has an FPGA. Due to the
fact that openbmc-linux has to be used, a patch is needed from Xilinx
Linux. The device tree for the BMC includes files that are generated
by the program that is used to program the FPGA (Vivado) which have to
be supplied by the user.

Tested: I was able to boot the generated binaries (uImage + devicetree +
initramfs) on the Mars ZX3 PM3 evaluation board and I was able to flash
the LEDs via the phosphor led manager.

Change-Id: I1eb7a17e6070dd8ea5ac312f9ea828e4cd389f50
Signed-off-by: Alexander Hedges <ahedges@ethz.ch>